==> 01.Data_Etl/03.TIMS_accident_prep.ipynb <==
# Prepare Transportation Injury Mapping System (TIMS) Data
In this notebook we download the collision data for Los Angeles County which has data for years 2014 to 2020.

%matplotlib inline

import pandas as pd
import numpy as np
from h3 import h3
import os
from os.path import isfile, join
from pathlib import Path
from os import listdir

import boto3
import awswrangler
# set name of S3 bucket
s3_bucket = 'traffic-data-bucket'

path =  Path(os.getcwd())
root = path.parent.absolute()

import warnings
pd.options.mode.chained_assignment = None  # default='warn'

h3_level = 8

## 1. Connect to AWS Services

from aws_secrets import aws_access_key_id, aws_secret_access_key, aws_session_token

my_session = boto3.Session(
    aws_access_key_id=aws_access_key_id,
    aws_secret_access_key=aws_secret_access_key,
    aws_session_token = aws_session_token

)

## 2. Download collision data from S3 bucket using AWS Wrangler
AWS Wrangler is used to read all files in the S3 Bucket with a .csv suffix into a single Pandas dataframe.

raw_path_dir = 'TIMS_raw_crashes/'

# path of S3 bucket where collision data is stored
raw_path = f"s3://{s3_bucket}/{raw_path_dir}"

# read data from S3 bucket
collision_df = awswrangler.s3.read_csv(path=raw_path, path_suffix=['.csv'], dataset=True,
                                 boto3_session=my_session, use_threads=True, low_memory=False)

collision_df.shape

collision_df.head()

## 3. Clean and create time features
The dataset contains some invalid times which will be dropped and from the `collission_date` variable we will extract the year, month, day, day of the week, and day of the year.

# 0:0>4 is a left padding of a string so if the string length is less than 4, add front ‘0’, e.g., 23 would become ‘0023’ 
collision_df['COLLISION_TIME']=collision_df['COLLISION_TIME'].apply(lambda x: '{0:0>4}'.format(x))

# create a mask to remove records with invalid times
# there are several rows that have '2500' entered as their collision time
mask = collision_df['COLLISION_TIME']=='2500'

# apply mask to capture records with valid times
valid_time = collision_df[~mask]

# create a dataframe of records with nonvalid times
nonvalid_time = collision_df[mask]
nonvalid_time['COLLISION_HOUR'] = 'none'
# print number of records to drop
print(nonvalid_time.shape)

valid_time = valid_time.assign(COLLISION_HOUR=pd.to_datetime(valid_time.COLLISION_TIME, format='%H%M').dt.hour)
valid_time = valid_time.assign(COLLISION_DATE=pd.to_datetime(valid_time.COLLISION_DATE, format='%Y-%m-%d'))

# create date features
valid_time['COLLISION_YEAR'] = valid_time.COLLISION_DATE.dt.year
valid_time['COLLISION_MONTH'] = valid_time.COLLISION_DATE.dt.month
valid_time['COLLISION_DAY'] = valid_time.COLLISION_DATE.dt.day
valid_time['COLLISION_DAYOFWEEK'] = valid_time.COLLISION_DATE.dt.dayofweek
valid_time['COLLISION_DAYOFYEAR'] = valid_time.COLLISION_DATE.dt.dayofyear

valid_time.head(2)

Next, lowercase each column name.

valid_time.columns = [each_string.lower() for each_string in valid_time.columns]

## 4. Create and attach hexagons using H3 library
`h3_level` refers to the size of each hexagon.  A higher value represents a larger hexagon which covers a larger area and smaller values represent smaller hexagons.

print('H3 Level:', h3_level)

# create a function to convert latitude and longitude values to hexagons
def lat_lng_to_h3(row):
    return h3.geo_to_h3(row.point_y, row.point_x, h3_level)

pd_dict = {}

valid_time['hex_id'] = valid_time.apply(lat_lng_to_h3, axis=1)
print(valid_time.shape)
valid_time = valid_time.groupby(['hex_id', 'collision_date', 'collision_year', 'collision_dayofyear', 'collision_month', 'collision_dayofweek', 'collision_hour']).first()
valid_time.reset_index(inplace = True)

valid_time_select = valid_time[['hex_id', 'collision_date', 'collision_year', 'collision_dayofyear', 'collision_month', 'collision_dayofweek', 'collision_hour']]
valid_time_select['accident_count'] = 1
print(valid_time_select.shape)
#valid_time_cnt.sample(3)
#valid_time_cnt.columns = ['h3_'+str(level), 'collision_count_h3_'+str(level)]
#pd_dict[key_val] = valid_time_cnt

## 5. Upload processed data to S3 Bucket
Convert the processed data to CSV and upload to the S3 bucket.

# upload to a new folder in the bucket called h3_processed_data
awswrangler.s3.to_csv(df=valid_time_select, path = f's3://{s3_bucket}/h3_processed_data/collisions_hex.csv', index=False,
                       boto3_session=my_session, use_threads=True
                       )
# upload to root of S3 Bucket
awswrangler.s3.to_csv(df=valid_time_select, path = f's3://{s3_bucket}/all_collisions_points.csv', index=False,
                       boto3_session=my_session, use_threads=True
                       )

### 5.1 Store local copy of sample data

valid_time_select[:50].to_csv(root / 'X.data' / 'all_collisions_points.csv', index=False)
